belafonte-ranker-master/requirements.txt
--index-url https://artifactory.spotify.net/artifactory/api/pypi/pypi/simple

# Packages for Luigi
spotify-scala-luigi==0.0.59
spotify-bigquery-luigi==0.1.6

# Add your own dependencies here
google-cloud-bigquery==1.23.1
google-cloud-core==1.1.0
google-auth-httplib2==0.0.3
google-cloud-storage==1.24.1

bqt==0.3.23
numpy==1.18.1
pandas==0.25.3
scikit-learn==0.21.2
scipy==1.4.1
xgboost==0.90
